requirements.txt
# Core dependencies
click>=8.0.0
pandas>=1.3.0
numpy>=1.20.0
duckdb>=0.7.0
zstandard>=0.18.0
pyyaml>=6.0
tabulate>=0.9.0

# For Parquet support
pyarrow>=7.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.950
flake8>=5.0.0
